Add CSR format to BiGraph and start parlmbc
In this commit, we add the feature to read the graph data with
the CSR format. Also, we test the correctness of CSR format
in CSR_unit_test folder

The parlmbc function, which is the parallel version of minelmbc
is also done, except for some bugs. (With the thread number
increases, the biclique number decreases. It is correct to use
single thread to execute.)